requirements.txt
ghapi
google-generativeai